requirements.txt
streamlit
firecrawl-py
pandas
sqlalchemy
